4.0.0 prep, remove BrawlAPI
