+++ .env.example
@@ -1,0 +1,2 @@
+token = # Replace this with your API token
+base_url = # Proxy server to handle requests to API due to IP limitations
