+++ CONTRIBUTING.md
@@ -8,8 +8,8 @@
 6. Add the necessary tests in the `tests` folder
 7. Add the necessary documentation and docstrings
 8. Add the necessary points to `CHANGELOG.md`
-9. Fill up the `tests/.env` file with the suitable token(s)
-10. Run `tox` from the root folder and ensure the tests are configured correctly and they return OK. `ServerError` and `MaintenanceError` can be disregarded.
+9. Fill up the `.env` file
+10. Run `tox` from the root folder and ensure the tests are configured correctly and they return OK. `ServerError` can be disregarded.
 11. Open your PR
 
 Do not increment version numbers but update `CHANGELOG.md`
