--- MANIFEST.in
@@ -1,4 +1,0 @@
-include README.md
-include LICENSE
-include requirements.txt
-include brawlstats/constants.json
