+++ brawlstats/errors.py
@@ -1,25 +1,15 @@
 class RequestError(Exception):
     """The base class for all errors."""
 
-    def __init__(self, code, message, retry_after=None):
+    def __init__(self, code, message):
         pass
 
     def __str__(self):
         return self.message
 
 
-class Unauthorized(RequestError):
-    """Raised if your API Key is invalid or blocked."""
-
-    def __init__(self, code, url):
-        self.code = code
-        self.url = url
-        self.message = 'Your API Key is invalid or blocked.'
-        super().__init__(self.code, self.message)
-
-
 class Forbidden(RequestError):
-    """Raised if the IP using the token was not whitelisted."""
+    """Raised if your API Key is invalid."""
 
     def __init__(self, code, url, message):
         self.code = code
@@ -31,21 +21,25 @@ class Forbidden(RequestError):
 class NotFoundError(RequestError):
     """Raised if an invalid player tag or club tag has been passed."""
 
-    def __init__(self, code, invalid_chars=[]):
+    def __init__(self, code, **kwargs):
         self.code = code
-        self.message = 'An incorrect tag has been passed.\nInvalid Characters: ' + ', '.join(invalid_chars)
-        self.invalid_chars = invalid_chars
+        self.message = 'An incorrect tag has been passed.'
+        self.reason = kwargs.pop('reason', None)
+        self.invalid_chars = kwargs.pop('invalid_chars', [])
+        if self.reason:
+            self.message += '\nReason: {}'.format(self.reason)
+        elif self.invalid_chars:
+            self.message += 'Invalid characters: {}'.format(', '.join(self.invalid_chars))
         super().__init__(self.code, self.message)
 
 
 class RateLimitError(RequestError):
     """Raised when the rate limit is reached."""
-    def __init__(self, code, url, retry_after):
+    def __init__(self, code, url):
         self.code = code
         self.url = url
-        self.retry_after = retry_after
-        self.message = 'The rate limit has been reached.\nRetry after: {}s'.format(retry_after)
-        super().__init__(self.code, self.message, retry_after=self.retry_after)
+        self.message = 'The rate limit has been reached.'
+        super().__init__(self.code, self.message)
 
 
 class UnexpectedError(RequestError):
@@ -65,14 +59,4 @@ class ServerError(RequestError):
         self.code = code
         self.url = url
         self.message = 'The API is down. Please be patient and try again later.'
-        super().__init__(self.code, self.message)
-
-
-class MaintenanceError(RequestError):
-    """Raised if there is a maintenance break."""
-
-    def __init__(self, code, url):
-        self.code = code
-        self.url = url
-        self.message = 'There is currently a maintenance break. Please be patient and try again later.'
         super().__init__(self.code, self.message)
