+++ requirements.txt
@@ -1,4 +1,4 @@
-aiohttp>=3.5.2,<3.6.0
+aiohttp>=3.6.0,<3.7.0
 requests
 python-box
-cachetools==3.1.0
+cachetools>=3.1.0
