--- tests/.env.example
@@ -1,3 +1,0 @@
-unofficial_token = # Replace this with your BrawlAPI Token
-official_token = # Replace this with your Official API token
-official_api_url = # Proxy server to handle requests to API due to IP limitations
